refactor types pkg, remove version string parsers, add PortPrefix
