+++ modules/apps/27-interchain-accounts/types/account_test.go
@@ -17,8 +17,9 @@ import (
 var (
 	// TestOwnerAddress defines a reusable bech32 address for testing purposes
 	TestOwnerAddress = "cosmos17dtl0mjt3t77kpuhg2edqzjpszulwhgzuj9ljs"
+
 	// TestPortID defines a resuable port identifier for testing purposes
-	TestPortID, _ = types.GeneratePortID(TestOwnerAddress, "connection-0", "connection-0")
+	TestPortID, _ = types.NewControllerPortID(TestOwnerAddress)
 )
 
 type TypesTestSuite struct {
