+++ modules/apps/27-interchain-accounts/types/port_test.go
@@ -7,7 +7,7 @@ import (
 	ibctesting "github.com/cosmos/ibc-go/v3/testing"
 )
 
-func (suite *TypesTestSuite) TestGeneratePortID() {
+func (suite *TypesTestSuite) TestNewControllerPortID() {
 	var (
 		path  *ibctesting.Path
 		owner = TestOwnerAddress
@@ -22,32 +22,8 @@ func (suite *TypesTestSuite) TestGeneratePortID() {
 		{
 			"success",
 			func() {},
-			fmt.Sprint(types.VersionPrefix, types.Delimiter, "0", types.Delimiter, "0", types.Delimiter, TestOwnerAddress),
+			fmt.Sprint(types.PortPrefix, types.Delimiter, TestOwnerAddress),
 			true,
-		},
-		{
-			"success with non matching connection sequences",
-			func() {
-				path.EndpointA.ConnectionID = "connection-1"
-			},
-			fmt.Sprint(types.VersionPrefix, types.Delimiter, "1", types.Delimiter, "0", types.Delimiter, TestOwnerAddress),
-			true,
-		},
-		{
-			"invalid connectionID",
-			func() {
-				path.EndpointA.ConnectionID = "connection"
-			},
-			"",
-			false,
-		},
-		{
-			"invalid counterparty connectionID",
-			func() {
-				path.EndpointB.ConnectionID = "connection"
-			},
-			"",
-			false,
 		},
 		{
 			"invalid owner address",
@@ -69,7 +45,7 @@ func (suite *TypesTestSuite) TestGeneratePortID() {
 
 			tc.malleate() // malleate mutates test data
 
-			portID, err := types.GeneratePortID(owner, path.EndpointA.ConnectionID, path.EndpointB.ConnectionID)
+			portID, err := types.NewControllerPortID(owner)
 
 			if tc.expPass {
 				suite.Require().NoError(err, tc.name)
@@ -77,92 +53,6 @@ func (suite *TypesTestSuite) TestGeneratePortID() {
 			} else {
 				suite.Require().Error(err, tc.name)
 				suite.Require().Empty(portID)
-			}
-		})
-	}
-}
-
-func (suite *TypesTestSuite) TestParseControllerConnSequence() {
-
-	testCases := []struct {
-		name     string
-		portID   string
-		expValue uint64
-		expPass  bool
-	}{
-		{
-			"success",
-			TestPortID,
-			0,
-			true,
-		},
-		{
-			"failed to parse port identifier",
-			"invalid-port-id",
-			0,
-			false,
-		},
-		{
-			"failed to parse connection sequence",
-			"ics27-1.x.y.cosmos1",
-			0,
-			false,
-		},
-	}
-
-	for _, tc := range testCases {
-		suite.Run(tc.name, func() {
-			connSeq, err := types.ParseControllerConnSequence(tc.portID)
-
-			if tc.expPass {
-				suite.Require().Equal(tc.expValue, connSeq)
-				suite.Require().NoError(err, tc.name)
-			} else {
-				suite.Require().Zero(connSeq)
-				suite.Require().Error(err, tc.name)
-			}
-		})
-	}
-}
-
-func (suite *TypesTestSuite) TestParseHostConnSequence() {
-
-	testCases := []struct {
-		name     string
-		portID   string
-		expValue uint64
-		expPass  bool
-	}{
-		{
-			"success",
-			TestPortID,
-			0,
-			true,
-		},
-		{
-			"failed to parse port identifier",
-			"invalid-port-id",
-			0,
-			false,
-		},
-		{
-			"failed to parse connection sequence",
-			"ics27-1.x.y.cosmos1",
-			0,
-			false,
-		},
-	}
-
-	for _, tc := range testCases {
-		suite.Run(tc.name, func() {
-			connSeq, err := types.ParseHostConnSequence(tc.portID)
-
-			if tc.expPass {
-				suite.Require().Equal(tc.expValue, connSeq)
-				suite.Require().NoError(err, tc.name)
-			} else {
-				suite.Require().Zero(connSeq)
-				suite.Require().Error(err, tc.name)
 			}
 		})
 	}
